+++ src/RcppExports.cpp
@@ -13,23 +13,38 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // calc_sum_squares_latent
-NumericVector calc_sum_squares_latent(arma::sp_mat Y, arma::mat X, arma::mat W, arma::vec ybar, int threads);
+NumericVector calc_sum_squares_latent(const arma::sp_mat& Y, const arma::mat& X, const arma::mat& W, const arma::vec& ybar, int threads);
 RcppExport SEXP _mvrsquared_calc_sum_squares_latent(SEXP YSEXP, SEXP XSEXP, SEXP WSEXP, SEXP ybarSEXP, SEXP threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::sp_mat >::type Y(YSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type W(WSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type ybar(ybarSEXP);
+    Rcpp::traits::input_parameter< const arma::sp_mat& >::type Y(YSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type X(XSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type W(WSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type ybar(ybarSEXP);
     Rcpp::traits::input_parameter< int >::type threads(threadsSEXP);
     rcpp_result_gen = Rcpp::wrap(calc_sum_squares_latent(Y, X, W, ybar, threads));
+    return rcpp_result_gen;
+END_RCPP
+}
+// calc_sum_squares
+NumericVector calc_sum_squares(const arma::sp_mat& Y, const arma::mat& Yhat, const arma::vec& ybar, int threads);
+RcppExport SEXP _mvrsquared_calc_sum_squares(SEXP YSEXP, SEXP YhatSEXP, SEXP ybarSEXP, SEXP threadsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::sp_mat& >::type Y(YSEXP);
+    Rcpp::traits::input_parameter< const arma::mat& >::type Yhat(YhatSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type ybar(ybarSEXP);
+    Rcpp::traits::input_parameter< int >::type threads(threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(calc_sum_squares(Y, Yhat, ybar, threads));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
     {"_mvrsquared_calc_sum_squares_latent", (DL_FUNC) &_mvrsquared_calc_sum_squares_latent, 5},
+    {"_mvrsquared_calc_sum_squares", (DL_FUNC) &_mvrsquared_calc_sum_squares, 4},
     {NULL, NULL, 0}
 };